requirements.txt
langchain
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
langchain-groq
